# docs/postman/servicio-academico-openapi.json
{"openapi":"3.0.1","info":{"title":"Servicio Académico API","description":"API para gestión de estudiantes, profesores, apoderados, asignaturas, evaluaciones y notas del Colegio Bernardo O'Higgins","contact":{"name":"Colegio Bernardo O'Higgins","email":"contacto@colegiobernardoohiggins.cl"},"license":{"name":"© 2025 Colegio Bernardo O'Higgins"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Notas","description":"Gestión de calificaciones"},{"name":"Estudiantes","description":"Gestión de estudiantes del colegio"},{"name":"Horarios","description":"Gestión de horarios de clases"},{"name":"Apoderados","description":"Gestión de apoderados y tutores"},{"name":"Asignaturas","description":"Gestión de asignaturas o materias"},{"name":"Evaluaciones","description":"Gestión de evaluaciones académicas"},{"name":"Autenticación Académico","description":"Validación de credenciales contra la base de datos"},{"name":"Profesores","description":"Gestión de profesores del colegio"}],"paths":{"/api/profesores/{id}":{"get":{"tags":["Profesores"],"summary":"Buscar profesor por ID","operationId":"buscar","parameters":[{"name":"id","in":"path","description":"ID del profesor","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Profesor no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profesor"}}}},"200":{"description":"Profesor encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profesor"}}}}}},"put":{"tags":["Profesores"],"summary":"Actualizar profesor","operationId":"actualizar","parameters":[{"name":"id","in":"path","description":"ID del profesor","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Profesor"}}},"required":true},"responses":{"404":{"description":"Profesor no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profesor"}}}},"200":{"description":"Profesor actualizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profesor"}}}}}},"delete":{"tags":["Profesores"],"summary":"Eliminar profesor","operationId":"eliminar","parameters":[{"name":"id","in":"path","description":"ID del profesor","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Profesor no encontrado"},"204":{"description":"Profesor eliminado"}}}},"/api/horarios/{id}":{"get":{"tags":["Horarios"],"summary":"Buscar horario por ID","description":"Obtiene un horario específico por su ID","operationId":"buscar_2","parameters":[{"name":"id","in":"path","description":"ID del horario","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Horario encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Horario"}}}},"404":{"description":"Horario no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Horario"}}}}}},"put":{"tags":["Horarios"],"summary":"Actualizar horario","description":"Actualiza los datos de un horario existente","operationId":"actualizar_1","parameters":[{"name":"id","in":"path","description":"ID del horario a actualizar","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"description":"Datos actualizados del horario","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Horario"}}},"required":true},"responses":{"404":{"description":"Horario no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Horario"}}}},"200":{"description":"Horario actualizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Horario"}}}}}},"delete":{"tags":["Horarios"],"summary":"Eliminar horario","description":"Elimina un horario del sistema","operationId":"eliminar_2","parameters":[{"name":"id","in":"path","description":"ID del horario a eliminar","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Horario no encontrado"},"204":{"description":"Horario eliminado"}}}},"/api/evaluaciones/{id}":{"get":{"tags":["Evaluaciones"],"summary":"Buscar evaluación por ID","operationId":"buscar_3","parameters":[{"name":"id","in":"path","description":"ID de la evaluación","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Evaluación encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Evaluacion"}}}},"404":{"description":"Evaluación no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Evaluacion"}}}}}},"put":{"tags":["Evaluaciones"],"summary":"Actualizar evaluación","operationId":"actualizar_2","parameters":[{"name":"id","in":"path","description":"ID de la evaluación","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Evaluacion"}}},"required":true},"responses":{"200":{"description":"Evaluación actualizada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Evaluacion"}}}},"404":{"description":"Evaluación no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Evaluacion"}}}}}},"delete":{"tags":["Evaluaciones"],"summary":"Eliminar evaluación","operationId":"eliminar_3","parameters":[{"name":"id","in":"path","description":"ID de la evaluación","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Evaluación eliminada"},"404":{"description":"Evaluación no encontrada"}}}},"/api/estudiantes/{id}":{"get":{"tags":["Estudiantes"],"summary":"Buscar estudiante por ID","description":"Obtiene los datos de un estudiante específico","operationId":"buscar_4","parameters":[{"name":"id","in":"path","description":"ID del estudiante","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Estudiante no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Estudiante"}}}},"200":{"description":"Estudiante encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Estudiante"}}}}}},"put":{"tags":["Estudiantes"],"summary":"Actualizar estudiante","description":"Actualiza los datos de un estudiante existente","operationId":"actualizar_3","parameters":[{"name":"id","in":"path","description":"ID del estudiante","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Estudiante"}}},"required":true},"responses":{"404":{"description":"Estudiante no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Estudiante"}}}},"200":{"description":"Estudiante actualizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Estudiante"}}}}}},"delete":{"tags":["Estudiantes"],"summary":"Eliminar estudiante","description":"Elimina un estudiante del sistema","operationId":"eliminar_4","parameters":[{"name":"id","in":"path","description":"ID del estudiante","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Estudiante eliminado"},"404":{"description":"Estudiante no encontrado"}}}},"/api/asignaturas/{id}":{"get":{"tags":["Asignaturas"],"summary":"Buscar asignatura por ID","operationId":"buscar_5","parameters":[{"name":"id","in":"path","description":"ID de la asignatura","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Asignatura encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Asignatura"}}}},"404":{"description":"Asignatura no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Asignatura"}}}}}},"put":{"tags":["Asignaturas"],"summary":"Actualizar asignatura","operationId":"actualizar_4","parameters":[{"name":"id","in":"path","description":"ID de la asignatura","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asignatura"}}},"required":true},"responses":{"200":{"description":"Asignatura actualizada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Asignatura"}}}},"404":{"description":"Asignatura no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Asignatura"}}}}}},"delete":{"tags":["Asignaturas"],"summary":"Eliminar asignatura","operationId":"eliminar_5","parameters":[{"name":"id","in":"path","description":"ID de la asignatura","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Asignatura eliminada"},"404":{"description":"Asignatura no encontrada"}}}},"/api/apoderados/{id}":{"get":{"tags":["Apoderados"],"summary":"Buscar apoderado por ID","operationId":"buscar_6","parameters":[{"name":"id","in":"path","description":"ID del apoderado","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Apoderado no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Apoderado"}}}},"200":{"description":"Apoderado encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Apoderado"}}}}}},"put":{"tags":["Apoderados"],"summary":"Actualizar apoderado","operationId":"actualizar_5","parameters":[{"name":"id","in":"path","description":"ID del apoderado","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Apoderado"}}},"required":true},"responses":{"404":{"description":"Apoderado no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Apoderado"}}}},"200":{"description":"Apoderado actualizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Apoderado"}}}}}},"delete":{"tags":["Apoderados"],"summary":"Eliminar apoderado","operationId":"eliminar_6","parameters":[{"name":"id","in":"path","description":"ID del apoderado","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Apoderado no encontrado"},"204":{"description":"Apoderado eliminado"}}}},"/api/profesores":{"get":{"tags":["Profesores"],"summary":"Listar todos los profesores","operationId":"listar","responses":{"200":{"description":"Lista de profesores","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Profesor"}}}}}}},"post":{"tags":["Profesores"],"summary":"Crear profesor","operationId":"crear","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Profesor"}}},"required":true},"responses":{"200":{"description":"Profesor creado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profesor"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Profesor"}}}}}}},"/api/notas":{"get":{"tags":["Notas"],"summary":"Listar todas las notas","operationId":"listar_1","responses":{"200":{"description":"Lista de notas","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Nota"}}}}}}},"post":{"tags":["Notas"],"summary":"Registrar nota","description":"Registra una nueva calificación para un estudiante","operationId":"crear_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Nota"}}},"required":true},"responses":{"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Nota"}}}},"200":{"description":"Nota registrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Nota"}}}}}}},"/api/horarios":{"get":{"tags":["Horarios"],"summary":"Listar todos los horarios","description":"Obtiene todos los horarios de clases registrados en el sistema","operationId":"listar_2","responses":{"200":{"description":"Lista de horarios","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Horario"}}}}}}},"post":{"tags":["Horarios"],"summary":"Crear horario","description":"Registra un nuevo bloque horario en el sistema","operationId":"crear_2","requestBody":{"description":"Datos del horario a crear","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Horario"}}},"required":true},"responses":{"200":{"description":"Horario creado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Horario"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Horario"}}}}}}},"/api/evaluaciones":{"get":{"tags":["Evaluaciones"],"summary":"Listar todas las evaluaciones","operationId":"listar_3","responses":{"200":{"description":"Lista de evaluaciones","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Evaluacion"}}}}}}},"post":{"tags":["Evaluaciones"],"summary":"Crear evaluación","operationId":"crear_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Evaluacion"}}},"required":true},"responses":{"200":{"description":"Evaluación creada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Evaluacion"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Evaluacion"}}}}}}},"/api/estudiantes":{"get":{"tags":["Estudiantes"],"summary":"Listar todos los estudiantes","description":"Obtiene la lista completa de estudiantes registrados","operationId":"listar_4","responses":{"200":{"description":"Lista de estudiantes","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Estudiante"}}}}}}},"post":{"tags":["Estudiantes"],"summary":"Crear estudiante","description":"Registra un nuevo estudiante en el sistema","operationId":"crear_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Estudiante"}}},"required":true},"responses":{"200":{"description":"Estudiante creado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Estudiante"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Estudiante"}}}}}}},"/api/auth/validar":{"post":{"tags":["Autenticación Académico"],"summary":"Validar credenciales de un usuario","description":"Busca el RUT en Profesores, Estudiantes y Apoderados, y verifica la contraseña con BCrypt","operationId":"validar","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/auth/resetear-contrasena":{"post":{"tags":["Autenticación Académico"],"summary":"Resetear contraseña de un usuario (solo UTP)","description":"Permite a un administrador UTP resetear la contraseña de cualquier usuario sin verificar la actual.","operationId":"resetearContrasena","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/auth/cambiar-contrasena":{"post":{"tags":["Autenticación Académico"],"summary":"Cambiar contraseña de un usuario","description":"Verifica la contraseña actual y la actualiza por una nueva. Aplica a Profesores, Estudiantes y Apoderados.","operationId":"cambiarContrasena","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/asignaturas":{"get":{"tags":["Asignaturas"],"summary":"Listar todas las asignaturas","operationId":"listar_5","responses":{"200":{"description":"Lista de asignaturas","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Asignatura"}}}}}}},"post":{"tags":["Asignaturas"],"summary":"Crear asignatura","operationId":"crear_5","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asignatura"}}},"required":true},"responses":{"200":{"description":"Asignatura creada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Asignatura"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Asignatura"}}}}}}},"/api/apoderados":{"get":{"tags":["Apoderados"],"summary":"Listar todos los apoderados","operationId":"listar_6","responses":{"200":{"description":"Lista de apoderados","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Apoderado"}}}}}}},"post":{"tags":["Apoderados"],"summary":"Crear apoderado","operationId":"crear_6","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Apoderado"}}},"required":true},"responses":{"200":{"description":"Apoderado creado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Apoderado"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Apoderado"}}}}}}},"/api/notas/{id}":{"get":{"tags":["Notas"],"summary":"Buscar nota por ID","operationId":"buscar_1","parameters":[{"name":"id","in":"path","description":"ID de la nota","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Nota no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Nota"}}}},"200":{"description":"Nota encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Nota"}}}}}},"delete":{"tags":["Notas"],"summary":"Eliminar nota","operationId":"eliminar_1","parameters":[{"name":"id","in":"path","description":"ID de la nota","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Nota no encontrada"},"204":{"description":"Nota eliminada"}}}},"/api/notas/estudiante/{estudianteId}":{"get":{"tags":["Notas"],"summary":"Buscar notas por estudiante","description":"Obtiene todas las notas de un estudiante específico","operationId":"buscarPorEstudiante","parameters":[{"name":"estudianteId","in":"path","description":"ID del estudiante","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Lista de notas del estudiante","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Nota"}}}}}}}},"/api/horarios/profesor/{profesorId}":{"get":{"tags":["Horarios"],"summary":"Buscar horarios por profesor","description":"Obtiene los horarios asignados a un profesor","operationId":"buscarPorProfesor","parameters":[{"name":"profesorId","in":"path","description":"ID del profesor","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Lista de horarios del profesor","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Horario"}}}}}}}},"/api/horarios/curso/{curso}":{"get":{"tags":["Horarios"],"summary":"Buscar horarios por curso","description":"Obtiene todos los horarios de un curso específico","operationId":"buscarPorCurso","parameters":[{"name":"curso","in":"path","description":"Nivel del curso (1-12)","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Lista de horarios del curso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Horario"}}}}}}}},"/api/horarios/curso/{curso}/dia/{dia}":{"get":{"tags":["Horarios"],"summary":"Buscar horarios por curso y día","description":"Obtiene los horarios de un curso en un día específico","operationId":"buscarPorCursoYDia","parameters":[{"name":"curso","in":"path","description":"Nivel del curso","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"dia","in":"path","description":"Día de la semana (1=Lunes...5=Viernes)","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Lista de horarios filtrada","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Horario"}}}}}}}},"/api/estudiantes/rut/{rut}":{"get":{"tags":["Estudiantes"],"summary":"Buscar estudiante por RUT","description":"Obtiene los datos de un estudiante usando su RUT","operationId":"buscarPorRut","parameters":[{"name":"rut","in":"path","description":"RUT del estudiante (ej: 12.345.678-9)","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Estudiante no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Estudiante"}}}},"200":{"description":"Estudiante encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Estudiante"}}}}}}},"/api/apoderados/rut/{rut}":{"get":{"tags":["Apoderados"],"summary":"Buscar apoderado por RUT","operationId":"buscarPorRut_1","parameters":[{"name":"rut","in":"path","description":"RUT del apoderado","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Apoderado no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Apoderado"}}}},"200":{"description":"Apoderado encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Apoderado"}}}}}}},"/api/apoderados/estudiante/{estudianteId}":{"get":{"tags":["Apoderados"],"summary":"Buscar apoderados por estudiante","description":"Obtiene los apoderados asociados a un estudiante","operationId":"buscarPorEstudiante_1","parameters":[{"name":"estudianteId","in":"path","description":"ID del estudiante","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Lista de apoderados del estudiante","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Apoderado"}}}}}}}}},"components":{"schemas":{"Profesor":{"required":["apellido","contrasena","nombre"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"apellido":{"type":"string"},"rut":{"type":"string"},"especialidad":{"type":"string"},"email":{"type":"string"},"contrasena":{"type":"string"}}},"Asignatura":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"nivelCurso":{"type":"integer","format":"int32"},"profesor":{"$ref":"#/components/schemas/Profesor"}}},"Horario":{"required":["curso","dia","horaFin","horaInicio"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"asignatura":{"$ref":"#/components/schemas/Asignatura"},"profesor":{"$ref":"#/components/schemas/Profesor"},"curso":{"type":"integer","format":"int32"},"dia":{"type":"integer","format":"int32"},"horaInicio":{"type":"string"},"horaFin":{"type":"string"},"sala":{"type":"string"}}},"Evaluacion":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"fecha":{"type":"string","format":"date"},"tipo":{"type":"string","enum":["PRUEBA","TAREA","TRABAJO","DISERTACION"]},"asignatura":{"$ref":"#/components/schemas/Asignatura"}}},"Estudiante":{"required":["apellido","nombre","password"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"apellido":{"type":"string"},"rut":{"type":"string"},"email":{"type":"string"},"curso":{"type":"integer","format":"int32"},"password":{"type":"string"}}},"Apoderado":{"required":["apellido","contrasena","nombre"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"apellido":{"type":"string"},"rut":{"type":"string"},"email":{"type":"string"},"contrasena":{"type":"string"},"estudiante":{"$ref":"#/components/schemas/Estudiante"}}},"Nota":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"estudiante":{"$ref":"#/components/schemas/Estudiante"},"evaluacion":{"$ref":"#/components/schemas/Evaluacion"},"valor":{"type":"number","format":"double"}}}}}}